The command to get the Route is language dependant! German language added.
Also interface selection would fail if there is no StaticIP!
This is fixed!

WARNING:
I changed the registry key to get the Interface Information.
The old one didn't work for Vista.
The new one works at least for XP and Vista.
If the old key is needed for some older os, we need to check both!