Some more work getting the new scapy hooked in.
Added some new files that seem to be required, while commenting out
some code that does not seem to be required.  Not all scapy-related 
codepaths have been tested, but at least some of them work.

More work still needs to be done to verify that all the required 
files/classes are present.